+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ### Changelog
 
+* 2026/09/14: Reply with a friendly error when the app is installed at personal scope instead of a Teams channel/team - [@dblock](https://github.com/dblock).
 * Your contribution here.